move functions related to comments to appropriate directories
