+++ controllers/comments.controllers.js
@@ -1,4 +1,5 @@
-const { insertCommentByArticleId } = require("../models/comments.models")
+const { selectArticleById } = require("../models/articles.models");
+const { insertCommentByArticleId, selectArticleCommentsByArticleId } = require("../models/comments.models")
 
 module.exports = {
 
@@ -8,6 +9,20 @@ module.exports = {
     insertCommentByArticleId(article_id, req.body)
       .then((comment) => {
         res.status(201).send({ comment })
+      })
+      .catch(next);
+  },
+
+  getArticleCommentsByArticleId: function(req, res, next) {
+    const { article_id } = req.params;
+
+    //Check article exists before selecting comments
+    selectArticleById(article_id)
+      .then(() => {
+        return selectArticleCommentsByArticleId(article_id)
+      })
+      .then((comments) => {
+        res.status(200).send({ comments });
       })
       .catch(next);
   }
