+++ models/articles.models.js
@@ -1,15 +1,6 @@
 const db = require('../db/connection.js');
 
 module.exports = {
-
-  selectArticleCommentsByArticleId: function(article_id) {
-    return db.query(`
-    SELECT * FROM comments
-    WHERE article_id = $1
-    `, [article_id])
-      .then(({ rows }) => rows);
-  },
-
   
   selectArticles: function() {
     return db.query(`
