+++ models/comments.models.js
@@ -2,6 +2,14 @@ const db = require('../db/connection.js');
 
 module.exports = {
 
+  selectArticleCommentsByArticleId: function(article_id) {
+    return db.query(`
+    SELECT * FROM comments
+    WHERE article_id = $1
+    `, [article_id])
+      .then(({ rows }) => rows);
+  },
+
   insertCommentByArticleId(article_id, {username, body}) {
     return db.query(`
     INSERT INTO comments
